requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==4.2
django-cors-headers==4.9.0
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
frozenlist==1.8.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
magic-filter==1.0.12
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.11.0
yarl==1.22.0
zope.interface==8.1
